Revert "resolves an issue that causes a race condition"
This reverts commit 992b8012c4b0734388c6487c556f54d6151a8bea by partially
undoing some (but not all) of the changes made by that commit.

The problem this reversion addresses is that, as of 3.3.8, elasticdump no longer sets a
non-zero status code when a failure occrrs during the dump phase.


Specifically:

elasticdump --input https://nonentity.example/temp --output /tmp/temp.json
echo $?

reports 0 when it should report 1.

The apparent reason for this is that the callback passed in the mainline now returns
a function that calls process.exit(1) instead of performing the call itself.

Since no other changes were made to ensure that the function returned by the
call was actually called process.exit(1) was never called, meaning that
elasticdump could fail failed silently (at least as far as the exit code was concerned).

One part of the original change has been left as is, since it seemed sound, specifically:

https://github.com/taskrabbit/elasticsearch-dump/commit/992b8012c4b0734388c6487c556f54d6151a8bea#diff-f60b04ac83f174b93b06625be32a86f5

Signed-off-by: Jon Seymour <jon@wildducktheories.com>